break up tam tests
