+++ tamcolors/tests/__init__.py
@@ -1,7 +1,9 @@
 from . import tam
 from . import tam_tools
+from . import tam_io
 from . import all_tests
 
 __all__ = ("tam",
            "tam_tools",
+           "tam_io",
            "all_tests")
